+++ src/application/index.js
@@ -1,3 +1,3 @@
 // import('./examples/basic.js');
-// import('./examples/buttons.js');
+// import('./examples/button.js');
 import('./examples/sliders.js');
